Current state (WIP)

+++ GameDatabase/Entities/AiScoreDatum.cs
@@ -1,11 +1,10 @@
 ﻿using SharedProject.Enums;
-using SharedProject.Models;
 
 namespace GameDatabase.Entities;
 
 public class AiScoreDatum
 {
-    public ulong Baid { get; set; }
+    public uint Baid { get; set; }
 
     public uint SongId { get; set; }
