+++ GameDatabase/Entities/Credential.cs
@@ -2,10 +2,10 @@
 {
     public partial class Credential
     {
-        public ulong Baid { get; set; }
+        public uint Baid { get; set; }
         public string Password { get; set; } = null!;
         public string Salt { get; set; } = null!;
-        
+
         public virtual UserDatum? Ba { get; set; }
     }
 }
