+++ GameDatabase/Entities/SongBestDatum.cs
@@ -2,16 +2,16 @@
 
 namespace GameDatabase.Entities
 {
-	public partial class SongBestDatum
-	{
-		public ulong Baid { get; set; }
-		public uint SongId { get; set; }
-		public Difficulty Difficulty { get; set; }
-		public uint BestScore { get; set; }
-		public uint BestRate { get; set; }
-		public CrownType BestCrown { get; set; }
-		public ScoreRank BestScoreRank { get; set; }
+    public partial class SongBestDatum
+    {
+        public uint Baid { get; set; }
+        public uint SongId { get; set; }
+        public Difficulty Difficulty { get; set; }
+        public uint BestScore { get; set; }
+        public uint BestRate { get; set; }
+        public CrownType BestCrown { get; set; }
+        public ScoreRank BestScoreRank { get; set; }
 
-		public virtual UserDatum? Ba { get; set; }
-	}
+        public virtual UserDatum? Ba { get; set; }
+    }
 }
