+++ TaikoLocalServer/Services/DanScoreDatumService.cs
@@ -12,14 +12,14 @@ public class DanScoreDatumService : IDanScoreDatumService
         this.context = context;
     }
 
-    public async Task<List<DanScoreDatum>> GetDanScoreDataList(ulong baid, DanType danType)
+    public async Task<List<DanScoreDatum>> GetDanScoreDataList(uint baid, DanType danType)
     {
         return await context.DanScoreData.Where(datum => datum.Baid == baid && datum.DanType == danType)
             .Include(datum => datum.DanStageScoreData)
             .ToListAsync();
     }
 
-    public async Task<DanScoreDatum?> GetSingleDanScoreDatum(ulong baid, uint danId, DanType danType)
+    public async Task<DanScoreDatum?> GetSingleDanScoreDatum(uint baid, uint danId, DanType danType)
     {
         return await context.DanScoreData.Include(datum => datum.DanStageScoreData)
             .FirstOrDefaultAsync(datum => datum.Baid == baid &&
