+++ TaikoLocalServer/Services/Interfaces/ICardService.cs
@@ -5,13 +5,13 @@ namespace TaikoLocalServer.Services.Interfaces;
 
 public interface ICardService
 {
-	public Task<Card?> GetCardByAccessCode(string accessCode);
+    public Task<Card?> GetCardByAccessCode(string accessCode);
 
-	public ulong GetNextBaid();
+    public uint GetNextBaid();
 
-	public Task<List<User>> GetUsersFromCards();
+    public Task<List<User>> GetUsersFromCards();
 
-	public Task AddCard(Card card);
-	
-	public Task<bool> DeleteCard(string accessCode);
+    public Task AddCard(Card card);
+
+    public Task<bool> DeleteCard(string accessCode);
 }
