+++ TaikoLocalServer/Services/Interfaces/ISongPlayDatumService.cs
@@ -4,7 +4,7 @@ namespace TaikoLocalServer.Services.Interfaces;
 
 public interface ISongPlayDatumService
 {
-	public Task<List<SongPlayDatum>> GetSongPlayDatumByBaid(ulong baid);
+    public Task<List<SongPlayDatum>> GetSongPlayDatumByBaid(uint baid);
 
-	public Task AddSongPlayDatum(SongPlayDatum datum);
+    public Task AddSongPlayDatum(SongPlayDatum datum);
 }
